moved save related code to SaveProjectScenario from OpenSaveProjectScenario
