+++ src/project/internal/opensaveprojectscenario.h
@@ -25,22 +25,11 @@
 #include "iopensaveprojectscenario.h"
 
 #include "modularity/ioc.h"
-#include "iprojectconfiguration.h"
-#include "iprojectfilescontroller.h"
 #include "interactive/iinteractive.h"
-#include "interactive/iplatforminteractive.h"
-
-#include "cloud/musescorecom/imusescorecomservice.h"
-#include "cloud/audiocom/iaudiocomservice.h"
 
 namespace mu::project {
 class OpenSaveProjectScenario : public IOpenSaveProjectScenario, public muse::Contextable
 {
-    muse::GlobalInject<IProjectConfiguration> configuration;
-    muse::GlobalInject<muse::cloud::IMuseScoreComService> museScoreComService;
-    muse::GlobalInject<muse::cloud::IAudioComService> audioComService;
-    muse::GlobalInject<muse::IPlatformInteractive> platformInteractive;
-    muse::ContextInject<IProjectFilesController> projectFilesController = { this };
     muse::ContextInject<muse::IInteractive> interactive = { this };
 
 public:
@@ -49,35 +38,6 @@ public:
     {
     }
 
-    muse::RetVal<SaveLocation> askSaveLocation(INotationProjectPtr project, SaveMode mode,
-                                               SaveLocationType preselectedType = SaveLocationType::Undefined) const override;
-
-    muse::RetVal<muse::io::path_t> askLocalPath(INotationProjectPtr project, SaveMode mode) const override;
-    muse::RetVal<CloudProjectInfo> askCloudLocation(INotationProjectPtr project, SaveMode mode) const override;
-    muse::RetVal<CloudProjectInfo> askPublishLocation(INotationProjectPtr project) const override;
-    muse::RetVal<CloudAudioInfo> askShareAudioLocation(INotationProjectPtr project) const override;
-
-    bool warnBeforeSavingToExistingPubliclyVisibleCloudProject() const override;
-
     void showCloudOpenError(const muse::Ret& ret) const override;
-    muse::Ret showCloudSaveError(const muse::Ret& ret, const CloudProjectInfo& info, bool isPublishShare,
-                                 bool alreadyAttempted) const override;
-    muse::Ret showAudioCloudShareError(const muse::Ret& ret) const override;
-
-private:
-    muse::RetVal<muse::Val> ensureAuthorization(const QString& cloudeCode, bool publishingScore, const std::string& text) const;
-
-    muse::RetVal<SaveLocationType> saveLocationType() const;
-    muse::RetVal<SaveLocationType> askSaveLocationType() const;
-
-    /// \param isPublishShare:
-    ///     false -> this is part of a "Save to cloud" action
-    ///     true -> this is part of a "Publish" action
-    muse::RetVal<CloudProjectInfo> doAskCloudLocation(INotationProjectPtr project, SaveMode mode, bool isPublishShare) const;
-
-    bool warnBeforePublishing(bool isPublishShare, muse::cloud::Visibility visibility) const;
-
-    muse::Ret warnCloudNotAvailableForUploading(bool isPublishShare) const;
-    muse::Ret warnCloudNotAvailableForSharingAudio() const;
 };
 }
